PBIS replacement feature merge

+++ ocaml/xapi/helpers.ml
@@ -43,7 +43,7 @@ let log_exn_continue msg f x =
 
 type log_output = Always | Never | On_failure
 
-let call_script ?(log_output = Always) ?env script args =
+let call_script ?(log_output = Always) ?env ?stdin script args =
   let should_log_output_on_success, should_log_output_on_failure =
     match log_output with
     | Always ->
@@ -60,7 +60,14 @@ let call_script ?(log_output = Always) ?env script args =
     let env =
       match env with None -> [|"PATH=" ^ Sys.getenv "PATH"|] | Some env -> env
     in
-    let output, _ = Forkhelpers.execute_command_get_output ~env script args in
+    let output, _ =
+      match stdin with
+      | None ->
+          Forkhelpers.execute_command_get_output ~env script args
+      | Some stdin ->
+          Forkhelpers.execute_command_get_output_send_stdin ~env script args
+            stdin
+    in
     if should_log_output_on_success then
       debug "%s %s succeeded [ output = '%s' ]" script (String.concat " " args)
         output ;
@@ -2037,7 +2044,7 @@ module FileSys : sig
   (* bash-like interface for manipulating files *)
   type path = string
 
-  val rmrf : path -> unit
+  val rmrf : ?rm_top:bool -> path -> unit
 
   val mv : src:path -> dest:path -> unit
 
@@ -2047,18 +2054,18 @@ module FileSys : sig
 end = struct
   type path = string
 
-  let rmrf path =
+  let rmrf ?(rm_top = true) path =
     let ( // ) = Filename.concat in
-    let rec rm path =
+    let rec rm rm_top path =
       let st = Unix.lstat path in
       match st.Unix.st_kind with
       | Unix.S_DIR ->
-          Sys.readdir path |> Array.iter (fun file -> rm (path // file)) ;
-          Unix.rmdir path
+          Sys.readdir path |> Array.iter (fun file -> rm true (path // file)) ;
+          if rm_top then Unix.rmdir path
       | _ ->
           Unix.unlink path
     in
-    try rm path
+    try rm rm_top path
     with e ->
       error "failed to remove %s" path ;
       raise e
